move navigation folder into common folder
