+++ src/App.vue
@@ -11,7 +11,7 @@
 
 <script lang="ts">
     import LoginModal from "@/components/modals/LoginModal.vue";
-    import Navbar from "@/components/navigation/navbar/Navbar.vue";
+    import Navbar from "@/components/common/navigation/navbar/Navbar.vue";
     import { store, useStore } from "@/store/store";
     import { ref, onMounted } from "vue";
     import { MutationTypes } from "./store/mutation-types";
