+++ src/components/common/navigation/navbar/admin/Navbar.vue
@@ -13,8 +13,8 @@
 </template>
 
 <script lang="ts">
-    import BaseNavbar from "@/components/navigation/navbar/common/BaseNavbar.vue";
-    import ManageAccountsMenu from "@/components/navigation/navbar/admin/administration/ManageAccountsMenu.vue";
+    import BaseNavbar from "../common/BaseNavbar.vue";
+    import ManageAccountsMenu from "./administration/ManageAccountsMenu.vue";
     import ProfileMenu from "../common/profile/ProfileMenu.vue";
 
     export default {
