+++ src/components/common/navigation/navbar/common/BaseMenu.vue
