+++ src/components/common/navigation/navbar/common/profile/ProfileMenu.vue
@@ -18,15 +18,15 @@
 </template>
 
 <script lang="ts">
-    import BaseMenu from "@/components/navigation/navbar/common/BaseMenu.vue";
-    import MenuBody from "@/components/navigation/navbar/common/profile/ProfileMenuBody.vue";
-    import { useStore } from "../../../../../store/store";
+    import BaseMenu from "../BaseMenu.vue";
+    import MenuBody from "../profile/ProfileMenuBody.vue";
+    import { useStore } from "@/store/store";
     import { ref } from "vue";
-    import User from "../../../../../api/api_models/user_management/User";
-    import Lecturer from "../../../../../api/api_models/user_management/Lecturer";
-    import Admin from "../../../../../api/api_models/user_management/Admin";
-    import Student from "../../../../../api/api_models/user_management/Student";
-    import { MutationTypes } from "../../../../../store/mutation-types";
+    import User from "@/api/api_models/user_management/User";
+    import Lecturer from "@/api/api_models/user_management/Lecturer";
+    import Admin from "@/api/api_models/user_management/Admin";
+    import Student from "@/api/api_models/user_management/Student";
+    import { MutationTypes } from "@/store/mutation-types";
 
     export default {
         components: {
