+++ src/components/common/navigation/navbar/lecturer/courses/CourseMenu.vue
@@ -7,7 +7,6 @@
                 <aside class="absolute bg-transparent md:w-48 left-0 lg:w-160 h-8" />
                 <div class="absolute left-0 lg:-left-16 lg:w-160 mt-4 z-30 lg:z-10 bg-white shadow-md rounded-lg overflow-hidden border">
                     <menu-body />
-                    <menu-footer />
                 </div>
             </div>
         </template>
@@ -15,15 +14,13 @@
 </template>
 
 <script lang="ts">
-    import BaseMenu from "@/components/navigation/navbar/common/BaseMenu.vue";
+    import BaseMenu from "../../common/BaseMenu.vue";
     import MenuBody from "./CourseMenuBody.vue";
-    import MenuFooter from "./CourseMenuFooter.vue";
 
     export default {
         components: {
             BaseMenu,
             MenuBody,
-            MenuFooter,
         },
     };
 </script>
