+++ src/components/common/navigation/navbar/lecturer/courses/CourseMenuBody.vue
