+++ src/components/common/navigation/navbar/student/Navbar.vue
@@ -13,9 +13,9 @@
 </template>
 
 <script lang="ts">
-    import BaseNavbar from "@/components/navigation/navbar/common/BaseNavbar.vue";
-    import CourseMenu from "@/components/navigation/navbar/student/courses/CourseMenu.vue";
-    import ProfileMenu from "@/components/navigation/navbar/common/profile/ProfileMenu.vue";
+    import BaseNavbar from "../common/BaseNavbar.vue";
+    import CourseMenu from "./courses/CourseMenu.vue";
+    import ProfileMenu from "../common/profile/ProfileMenu.vue";
 
     export default {
         name: "AdminNavbar",
