+++ src/components/common/navigation/navbar/student/courses/CourseMenuBody.vue
