+++ src/components/common/navigation/navbar/student/courses/CourseMenuFooter.vue
@@ -33,7 +33,7 @@
 </template>
 
 <script lang="ts">
-    import MenuItem from "@/components/navigation/navbar/common/MenuItem.vue";
+    import MenuItem from "../../common/MenuItem.vue";
 
     export default {
         name: "CourseMenuFooter",
